requirements.txt
discord.py>=1.6.0
apscheduler>=3.7.0
